# src/docs/swagger.yaml
basePath: /v1
definitions:
  api.ConfirmWithdrawalResponse:
    properties:
      operation_id:
        type: string
      status:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  api.WithdrawalRequest:
    properties:
      amount:
        example: 10
        type: integer
      currency:
        enum:
        - USDT
        example: USDT
        type: string
      destination:
        example: TJRyWwFtzThYgBstgex9NL4iuMMryZQkwy
        type: string
      user_id:
        example: 12345678-1234-4123-b234-123456789012
        type: string
    required:
    - amount
    - currency
    - destination
    - user_id
    type: object
  api.WithdrawalResponse:
    properties:
      idempotency_key:
        type: string
      user_id:
        type: string
      withdrawal_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for creating and managing crypto withdrawals.
  title: Crypto Withdrawal API
  version: "1.0"
paths:
  /withdrawals:
    post:
      consumes:
      - application/json
      description: 'Creates a new crypto withdrawal. Idempotent: repeated requests
        with the same key and body return the cached response.'
      parameters:
      - description: Idempotency key (UUID v4)
        example: 550e8400-e29b-41d4-a716-446655440000
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Withdrawal payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.WithdrawalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.WithdrawalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Insufficient balance
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Idempotency key reused with different body
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create withdrawal
      tags:
      - withdrawals
  /withdrawals/{id}:
    get:
      description: Returns all withdrawals for the given user.
      parameters:
      - description: User ID (UUID)
        example: 12345678-1234-4123-b234-123456789012
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.WithdrawalResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get withdrawals
      tags:
      - withdrawals
  /withdrawals/{id}/confirm:
    post:
      description: Transitions a pending withdrawal to the complete state and writes
        a ledger entry.
      parameters:
      - description: Operation ID (UUID, returned from POST /withdrawals)
        example: 550e8400-e29b-41d4-a716-446655440001
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ConfirmWithdrawalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Withdrawal not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Withdrawal is not in pending status
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Confirm withdrawal
      tags:
      - withdrawals
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"